requirements.txt
# Dependencias principales
Flask==2.3.3
Werkzeug==2.3.7
pytz==2023.3

# Para notificaciones por correo
Flask-Mail==0.9.1

# Para conexiones HTTP con API externa
requests==2.31.0
httpx==0.27.0

# Para redención automática de pines (Free Fire ID)
greenlet>=3.1.0
playwright>=1.49.0

# Utilidades (opcional)
python-dotenv==1.0.0

# Para producción (opcional)
gunicorn==21.2.0

# PostgreSQL (Python 3.13 compatible)
psycopg[binary]==3.2.1

# Nota: sqlite3 fue reemplazado por psycopg. secrets, hashlib y os son módulos integrados de Python
